% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymedia-package.R
\docType{package}
\name{shinymedia-package}
\alias{shinymedia}
\alias{shinymedia-package}
\title{shinymedia: Multimedia Input/Output Controls for Shiny}
\description{
Extends the "Shiny" web framework with controls for multimedia input/output.
}
\author{
\strong{Maintainer}: Joe Cheng \email{joe@posit.co}

}
\keyword{internal}
